# obai_fullstack_package-v2/docs/openapi/obai-openapi-v1.yaml
openapi: 3.0.3
info:
  title: OBAI Platform API
  version: 1.0.0
  description: 耦白 OBAI 核心接口摘要。完整接口以运行后的 /v3/api-docs 为准。
servers:
  - url: http://localhost:8080
paths:
...
  /api/admin/{endpoint}:
    get:
      summary: 管理端通用列表
      parameters:
        - in: path
          name: endpoint
          required: true
          schema: { type: string }
      responses:
        '200': { description: List data }
    post:
      summary: 管理端通用新增
      parameters:
        - in: path
          name: endpoint
          required: true
          schema: { type: string }
      responses:
        '200': { description: Created entity }
  /api/admin/{endpoint}/{id}:
    get:
      summary: 管理端通用详情
      parameters:
        - in: path
          name: endpoint
          required: true
          schema: { type: string }
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200': { description: Entity detail }
    put:
      summary: 管理端通用更新
      parameters:
        - in: path
          name: endpoint
          required: true
          schema: { type: string }
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200': { description: Updated entity }
    delete:
      summary: 管理端通用删除
      parameters:
        - in: path
          name: endpoint
          required: true
          schema: { type: string }
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200': { description: Deleted }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT